+++ src/Interfaces/order/Order.ts
@@ -1,5 +1,6 @@
 import User from '../user/User';
 import OrderLine from './OrderLine';
+import Assignee from './Assignee';
 
 export default interface Order {
     id?: string;
@@ -47,8 +48,10 @@ export default interface Order {
     deliveryMethod?: string;
     expectedDeliveryDate?: string;
     hasReusablePackagingOptIn?: boolean;
+    weight?: number;
     // belongs to
     user?: User | null;
+    assignee?: Assignee | null;
     // has many
     lines?: OrderLine[];
 }