+++ src/Interfaces/user/User.ts
@@ -1,7 +1,7 @@
 import Order from '../order/Order';
 import Stock from '../stock/Stock';
 import { Basket } from '../basket/Basket';
-import UserSubscription from './UserSubscription';
+import UserSubscription, { UserSubscriptionStatusId } from './UserSubscription';
 import Address from '../address/Address';
 import ReturnOrder from '../return/ReturnOrder';
 import PurchaseStock from '../stock/PurchaseStock';
@@ -42,6 +42,7 @@ export default interface User {
     verifiedAt?: Date;
     reference?: string;
     resourceType?: string;
+    color?: string;
     highlights?: [];
     // has many
     orders?: Array<Order>;
@@ -52,7 +53,9 @@ export default interface User {
     returnOrders?: Array<ReturnOrder>;
     activeBasket?: Array<Basket>;
     userSubscriptions?: Array<UserSubscription>;
-    activeUserSubscription?: UserSubscription;
+    activeUserSubscriptions?: Array<UserSubscription>;
+    activeUserSubscriptionStatusId?: UserSubscriptionStatusId;
+    activeUserSubscriptionPlan?: string;
     activeReturnOrder?: ReturnOrder;
     roles?: Array<Role>;
     // Accessors
@@ -60,10 +63,12 @@ export default interface User {
     toyboxTokens?: number;
     additionalTokens?: number;
     availableTokens?: number;
+    subscriptionTokens?: number;
     maxTokens?: number;
     basketQuantity?: number;
     activeBasketTotalTokens?: number;
     balance?: number;
     changeSubscriptionStatus?: ChangeSubscriptionStatus;
     cancelSubscriptionStatus?: CancelSubscriptionStatus;
+    notes?: string;
 }