+++ src/Models/order/Transformers/OrderTransformer.ts
@@ -5,6 +5,8 @@ import UserInterface from '../../../Interfaces/user/User';
 import UserTransformer from '../../user/Transformers/UserTransformer';
 import OrderLineTransformer from '../../order/Transformers/OrderLineTransformer';
 import BaseTransformer from '../../../BaseTransformer';
+import AssigneeInterface from '../../../Interfaces/order/Assignee';
+import AssigneeTransformer from './AssigneeTransformer';
 
 export default class OrderTransformer extends BaseTransformer {
     /**
@@ -59,6 +61,7 @@ export default class OrderTransformer extends BaseTransformer {
             deliveryMethod: order.deliveryMethod,
             expectedDeliveryDate: order.expectedDeliveryDate,
             hasReusablePackagingOptIn: order.hasReusablePackagingOptIn,
+            weight: order.weight,
             // belongs to
             user: this.includeUser(order),
             // has many
@@ -72,5 +75,9 @@ export default class OrderTransformer extends BaseTransformer {
 
     includeOrderLines(order: OrderInterface): OrderLineInterface[] {
         return this.collection(order, 'lines', new OrderLineTransformer());
+    }
+
+    includeAssignee(order: OrderInterface): AssigneeInterface | null {
+        return this.item(order, 'assignee', new AssigneeTransformer());
     }
 }